move models to folders
